+++ examples/app1/src/app/dashboard/dashboard.component.ts
@@ -3,7 +3,7 @@ import { Router } from '@angular/router';
 import { CounterService } from '../counter.service';
 import { AppRootContext } from '@demo/common';
 import { PlanetComponentRef, PlanetComponentLoader, PlanetPortalApplication, GlobalEventDispatcher } from 'ngx-planet';
-import { ThyDialog, ThyDialogSizes } from 'ngx-tethys';
+import { ThyDialog, ThyDialogSizes } from 'ngx-tethys/dialog';
 import { App1DetailComponent } from '../detail/detail.component';
 import { ProjectsDialogComponent } from '../projects/dialog/projects-dialog.component';