+++ packages/planet/src/application/planet-application-loader.ts
@@ -10,6 +10,8 @@ import { PlanetApplicationService } from './planet-application.service';
 import { GlobalEventDispatcher } from '../global-event-dispatcher';
 import { Router } from '@angular/router';
 import { globalPlanet, getPlanetApplicationRef, getApplicationLoader } from '../global-planet';
+import { createDebug } from '../debug';
+const debug = createDebug('app-loader');
 
 export enum ApplicationStatus {
     assetsLoading = 1,
@@ -93,12 +95,32 @@ export class PlanetApplicationLoader {
 
     private setAppStatus(app: PlanetApplication, status: ApplicationStatus) {
         this.ngZone.run(() => {
+            const fromStatus = this.appsStatus.get(app);
+            debug(
+                `app(${app.name}) status change: ${fromStatus ? ApplicationStatus[fromStatus] : 'empty'} => ${
+                    ApplicationStatus[status]
+                }`
+            );
+            this.appsStatus.set(app, status);
             this.appStatusChange$.next({
                 app: app,
                 status: status
             });
-            this.appsStatus.set(app, status);
         });
+    }
+
+    private getAppStatusChange$(
+        app: PlanetApplication,
+        status = ApplicationStatus.bootstrapped
+    ): Observable<PlanetApplication> {
+        return this.appStatusChange.pipe(
+            filter(event => {
+                return event.app === app && event.status === status;
+            }),
+            map(() => {
+                return app;
+            })
+        );
     }
 
     private switchModeIsCoexist(app: PlanetApplication) {
@@ -115,14 +137,18 @@ export class PlanetApplicationLoader {
         });
     }
 
-    private takeOneStable() {
-        return this.ngZone.onStable.pipe(take(1));
-    }
-
     private setLoadingDone() {
         this.ngZone.run(() => {
             this.loadingDone = true;
         });
+    }
+
+    private getAppNames(apps: PlanetApplication[]): string | string[] {
+        return apps.length === 0
+            ? `[]`
+            : apps.map(item => {
+                  return item.name;
+              });
     }
 
     private setupRouteChange() {
@@ -138,14 +164,17 @@ export class PlanetApplicationLoader {
                     return of(event).pipe(
                         // unload apps and return should load apps
                         map(() => {
+                            debug(`route change, url is: ${event.url}`);
                             this.startRouteChangeEvent = event;
                             const shouldLoadApps = this.planetApplicationService.getAppsByMatchedUrl(event.url);
+                            debug(`should load apps: ${this.getAppNames(shouldLoadApps)}`);
                             const shouldUnloadApps = this.getUnloadApps(shouldLoadApps);
                             this.appsLoadingStart$.next({
                                 shouldLoadApps,
                                 shouldUnloadApps
                             });
                             this.unloadApps(shouldUnloadApps, event);
+                            debug(`unload apps: ${this.getAppNames(shouldUnloadApps)}`);
                             return shouldLoadApps;
                         }),
                         // Load app assets (static resources)
@@ -158,6 +187,9 @@ export class PlanetApplicationLoader {
                                     appStatus === ApplicationStatus.assetsLoading ||
                                     appStatus === ApplicationStatus.loadError
                                 ) {
+                                    debug(
+                                        `app(${app.name}) status is ${ApplicationStatus[appStatus]}, start load assets`
+                                    );
                                     hasAppsNeedLoadingAssets = true;
                                     return this.ngZone.runOutsideAngular(() => {
                                         return this.startLoadAppAssets(app);
@@ -173,58 +205,65 @@ export class PlanetApplicationLoader {
                         }),
                         // Bootstrap or show apps
                         map(apps => {
-                            const apps$: Observable<PlanetApplication>[] = [];
-                            apps.forEach(app => {
-                                const appStatus = this.appsStatus.get(app);
-                                if (appStatus === ApplicationStatus.bootstrapped) {
-                                    apps$.push(
-                                        of(app).pipe(
-                                            tap(() => {
-                                                this.showApp(app);
-                                                const appRef = getPlanetApplicationRef(app.name);
+                            const apps$: Observable<PlanetApplication>[] = apps.map(app => {
+                                return of(app).pipe(
+                                    switchMap(app => {
+                                        const appStatus = this.appsStatus.get(app);
+                                        if (appStatus === ApplicationStatus.bootstrapped) {
+                                            debug(
+                                                `[routeChange] app(${app.name}) status is bootstrapped, show app and active`
+                                            );
+                                            this.showApp(app);
+                                            const appRef = getPlanetApplicationRef(app.name);
+                                            appRef.navigateByUrl(event.url);
+                                            this.setAppStatus(app, ApplicationStatus.active);
+                                            this.setLoadingDone();
+                                            return of(app);
+                                        } else if (appStatus === ApplicationStatus.assetsLoaded) {
+                                            debug(
+                                                `[routeChange] app(${app.name}) status is assetsLoaded, start bootstrapping`
+                                            );
+                                            return this.bootstrapApp(app).pipe(
+                                                map(() => {
+                                                    debug(`app(${app.name}) bootstrapped success, active it`);
+                                                    this.setAppStatus(app, ApplicationStatus.active);
+                                                    this.setLoadingDone();
+                                                    return app;
+                                                })
+                                            );
+                                        } else if (appStatus === ApplicationStatus.active) {
+                                            debug(`[routeChange] app(${app.name}) is active, do nothings`);
+                                            const appRef = getPlanetApplicationRef(app.name);
+                                            // Backwards compatibility sub app use old version which has not getCurrentRouterStateUrl
+                                            const currentUrl = appRef.getCurrentRouterStateUrl
+                                                ? appRef.getCurrentRouterStateUrl()
+                                                : '';
+                                            if (currentUrl !== event.url) {
                                                 appRef.navigateByUrl(event.url);
-                                                this.setAppStatus(app, ApplicationStatus.active);
-                                                this.setLoadingDone();
-                                            })
-                                        )
-                                    );
-                                } else if (appStatus === ApplicationStatus.assetsLoaded) {
-                                    apps$.push(
-                                        of(app).pipe(
-                                            switchMap(() => {
-                                                return this.bootstrapApp(app).pipe(
-                                                    map(() => {
-                                                        this.setAppStatus(app, ApplicationStatus.active);
-                                                        this.setLoadingDone();
-                                                        return app;
-                                                    })
-                                                );
-                                            })
-                                        )
-                                    );
-                                } else if (appStatus === ApplicationStatus.active) {
-                                    apps$.push(
-                                        of(app).pipe(
-                                            tap(() => {
-                                                const appRef = getPlanetApplicationRef(app.name);
-                                                // Backwards compatibility sub app use old version which has not getCurrentRouterStateUrl
-                                                const currentUrl = appRef.getCurrentRouterStateUrl
-                                                    ? appRef.getCurrentRouterStateUrl()
-                                                    : '';
-                                                if (currentUrl !== event.url) {
-                                                    appRef.navigateByUrl(event.url);
-                                                }
-                                            })
-                                        )
-                                    );
-                                } else {
-                                    throw new Error(
-                                        `app(${app.name})'s status is ${appStatus}, can't be show or bootstrap`
-                                    );
-                                }
+                                            }
+                                            return of(app);
+                                        } else {
+                                            debug(
+                                                `[routeChange] app(${app.name}) status is ${ApplicationStatus[appStatus]}`
+                                            );
+                                            return this.getAppStatusChange$(app).pipe(
+                                                take(1),
+                                                map(() => {
+                                                    debug(
+                                                        `app(${app.name}) status is bootstrapped by subscribe status change, active it`
+                                                    );
+                                                    this.setAppStatus(app, ApplicationStatus.active);
+                                                    this.showApp(app);
+                                                    return app;
+                                                })
+                                            );
+                                        }
+                                    })
+                                );
                             });
 
                             if (apps$.length > 0) {
+                                debug(`start load and active apps: ${this.getAppNames(apps)}`);
                                 // 切换到应用后会有闪烁现象，所以使用 setTimeout 后启动应用
                                 // example: redirect to app1's dashboard from portal's about page
                                 // If app's route has redirect, it doesn't work, it ok just in setTimeout, I don't know why.
@@ -242,12 +281,14 @@ export class PlanetApplicationLoader {
                                     }
                                 });
                             } else {
+                                debug(`no apps need to be loaded, ensure preload apps`);
                                 this.ensurePreloadApps(apps);
                                 this.setLoadingDone();
                             }
                         }),
                         // Error handler
                         catchError(error => {
+                            debug(`apps loader error: ${error}`);
                             this.errorHandler(error);
                             return [];
                         })
@@ -314,6 +355,7 @@ export class PlanetApplicationLoader {
         app: PlanetApplication,
         defaultStatus: 'hidden' | 'display' = 'display'
     ): Observable<PlanetApplicationRef> {
+        debug(`app(${app.name}) start bootstrapping`);
         this.setAppStatus(app, ApplicationStatus.bootstrapping);
         const appRef = getPlanetApplicationRef(app.name);
         if (appRef && appRef.bootstrap) {
@@ -344,6 +386,7 @@ export class PlanetApplicationLoader {
             }
             return result.pipe(
                 tap(() => {
+                    debug(`app(${app.name}) bootstrapped success for ${defaultStatus}`);
                     this.setAppStatus(app, ApplicationStatus.bootstrapped);
                     if (defaultStatus === 'display' && appRootElement) {
                         appRootElement.removeAttribute('style');
@@ -375,6 +418,7 @@ export class PlanetApplicationLoader {
         const destroyApps: PlanetApplication[] = [];
         shouldUnloadApps.forEach(app => {
             if (this.switchModeIsCoexist(app)) {
+                debug(`hide app(${app.name}) for coexist mode`);
                 hideApps.push(app);
                 this.hideApp(app);
                 this.setAppStatus(app, ApplicationStatus.bootstrapped);
@@ -398,6 +442,7 @@ export class PlanetApplicationLoader {
                     }
                 });
                 destroyApps.forEach(app => {
+                    debug(`destroy app(${app.name})`);
                     this.destroyApp(app);
                 });
             });
@@ -409,6 +454,7 @@ export class PlanetApplicationLoader {
             const toPreloadApps = this.planetApplicationService.getAppsToPreload(
                 activeApps ? activeApps.map(item => item.name) : null
             );
+            debug(`start preload apps: ${this.getAppNames(toPreloadApps)}`);
             const loadApps$ = toPreloadApps.map(preloadApp => {
                 return this.preloadInternal(preloadApp);
             });
@@ -445,8 +491,10 @@ export class PlanetApplicationLoader {
     private preloadInternal(app: PlanetApplication, immediate?: boolean): Observable<PlanetApplicationRef> {
         const status = this.appsStatus.get(app);
         if (!status || status === ApplicationStatus.loadError) {
+            debug(`preload app(${app.name}), status is empty, start to load assets`);
             return this.startLoadAppAssets(app).pipe(
                 switchMap(() => {
+                    debug(`preload app(${app.name}), assets loaded, start bootstrap app, immediate: ${!!immediate}`);
                     if (immediate) {
                         return this.bootstrapApp(app, 'hidden');
                     } else {
@@ -464,10 +512,8 @@ export class PlanetApplicationLoader {
                 status
             )
         ) {
-            return this.appStatusChange.pipe(
-                filter(event => {
-                    return event.app === app && event.status === ApplicationStatus.bootstrapped;
-                }),
+            debug(`preload app(${app.name}), status is ${ApplicationStatus[status]}, return until bootstrapped`);
+            return this.getAppStatusChange$(app).pipe(
                 take(1),
                 map(() => {
                     return getPlanetApplicationRef(app.name);