+++ packages/planet/src/assets-loader.spec.ts
@@ -1,4 +1,4 @@
-import { TestBed, async, fakeAsync } from '@angular/core/testing';
+import { TestBed, fakeAsync, waitForAsync } from '@angular/core/testing';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 import { AssetsLoader, AssetsLoadResult } from './assets-loader';
 import { hashCode } from './helpers';
@@ -23,7 +23,7 @@ describe('assets-loader', () => {
     describe('loadScript', () => {
         it('should load script success when not in IE', fakeAsync(() => {
             const src = 'assets/main.js';
-            const createElementSpy: jasmine.Spy<InferableFunction> = spyOn(document, 'createElement');
+            const createElementSpy: jasmine.Spy = spyOn(document, 'createElement');
             const appendChildSpy = spyOn(document.body, 'appendChild');
 
             // 返回 mock script
@@ -64,7 +64,7 @@ describe('assets-loader', () => {
                 }
             };
             const src = 'assets/main.js';
-            const createElementSpy: jasmine.Spy<InferableFunction> = spyOn(document, 'createElement');
+            const createElementSpy: jasmine.Spy = spyOn(document, 'createElement');
             const appendChildSpy = spyOn(document.body, 'appendChild');
 
             // 返回 mock script
@@ -107,7 +107,7 @@ describe('assets-loader', () => {
                 }
             };
             const src = 'assets/main.js';
-            const createElementSpy: jasmine.Spy<InferableFunction> = spyOn(document, 'createElement');
+            const createElementSpy: jasmine.Spy = spyOn(document, 'createElement');
             const appendChildSpy = spyOn(document.body, 'appendChild');
 
             // 返回 mock script
@@ -150,7 +150,7 @@ describe('assets-loader', () => {
                 }
             };
             const src = 'assets/main.js';
-            const createElementSpy: jasmine.Spy<InferableFunction> = spyOn(document, 'createElement');
+            const createElementSpy: jasmine.Spy = spyOn(document, 'createElement');
             const appendChildSpy = spyOn(document.body, 'appendChild');
 
             // 返回 mock script
@@ -181,7 +181,7 @@ describe('assets-loader', () => {
 
         it('should not load script which has been loaded', fakeAsync(() => {
             const src = 'assets/main.js';
-            const createElementSpy: jasmine.Spy<InferableFunction> = spyOn(document, 'createElement');
+            const createElementSpy: jasmine.Spy = spyOn(document, 'createElement');
             const appendChildSpy = spyOn(document.body, 'appendChild');
 
             createElementSpy.and.returnValue(mockScript);
@@ -209,7 +209,7 @@ describe('assets-loader', () => {
 
         it('should get error when load script fail', fakeAsync(() => {
             const src = 'assets/main.js';
-            const createElementSpy: jasmine.Spy<InferableFunction> = spyOn(document, 'createElement');
+            const createElementSpy: jasmine.Spy = spyOn(document, 'createElement');
             const appendChildSpy = spyOn(document.body, 'appendChild');
 
             // 返回 mock script
@@ -360,25 +360,28 @@ describe('assets-loader', () => {
             httpTestingController.verify();
         });
 
-        it('should load manifest success', async(() => {
-            const testData = {
-                'main.js': 'main1.js'
-            };
-            const loadManifestSpy = jasmine.createSpy('load manifest spy');
-            assetsLoader.loadManifest('/static/assets/manifest.json').subscribe(loadManifestSpy);
-            const req = httpTestingController.expectOne('/static/assets/manifest.json');
+        it(
+            'should load manifest success',
+            waitForAsync(() => {
+                const testData = {
+                    'main.js': 'main1.js'
+                };
+                const loadManifestSpy = jasmine.createSpy('load manifest spy');
+                assetsLoader.loadManifest('/static/assets/manifest.json').subscribe(loadManifestSpy);
+                const req = httpTestingController.expectOne('/static/assets/manifest.json');
 
-            // Assert that the request is a GET.
-            expect(req.request.method).toEqual('GET');
-            expect(loadManifestSpy).not.toHaveBeenCalled();
+                // Assert that the request is a GET.
+                expect(req.request.method).toEqual('GET');
+                expect(loadManifestSpy).not.toHaveBeenCalled();
 
-            // Respond with mock data, causing Observable to resolve.
-            // Subscribe callback asserts that correct data was returned.
-            req.flush(testData);
+                // Respond with mock data, causing Observable to resolve.
+                // Subscribe callback asserts that correct data was returned.
+                req.flush(testData);
 
-            expect(loadManifestSpy).toHaveBeenCalled();
-            expect(loadManifestSpy).toHaveBeenCalledWith(testData);
-        }));
+                expect(loadManifestSpy).toHaveBeenCalled();
+                expect(loadManifestSpy).toHaveBeenCalledWith(testData);
+            })
+        );
     });
 
     describe('loadAppAssets', () => {