+++ packages/planet/src/testing/utils.ts
@@ -1,3 +1,5 @@
+import { setDebugFactory, clearDebugFactory } from '../debug';
+
 export function roundNumber(minNum: number, maxNum?: number): number {
     switch (arguments.length) {
         case 1:
@@ -12,4 +14,18 @@ export function roundNumber(minNum: number, maxNum?: number): number {
 export function sample<T>(items: T[]): T {
     const num = roundNumber(1, items.length);
     return items[num - 1];
+}
+
+export function enableDebugForTest() {
+    const debugFactory = (namespace: string) => {
+        return (formatter: string) => {
+            console.log(`[${namespace}] ${formatter}`);
+        };
+    };
+
+    setDebugFactory(debugFactory as any);
+}
+
+export function disableDebugForTest() {
+    clearDebugFactory();
 }